Create basic reports page system
Also added a current members page, though that's also still very basic.
Hopefully this should be easy to expand on in the future.

A warning I've noticed is mention of creating multiple database objects, I think it's not caching/reusing the db object in db-manager properly. So that will need to be looked into