Add 'add new member' page
The api doesn't yet do anything, on submitting the form it will respond with 201 and always refer to member id 101054, but the page itself should work fine now.
Just beginning work on the edit member page now.

Also fixed a couple of bugs/warnings in keys in the large member lists.